battle system - draw
First stage of the battle system - drawing cards. I'm working to
architect this in a way that I could transition it to an actual
client / server setup later on which is why so much is spread out.

There's several TODOs and a lot of things I'm not quite happy with, but
I want to get the core game playable ASAP to keep motivation up.